[LEMS-4497/modern-sorter-editor] split out card-editor
